backend/requirements.txt
fastapi
uvicorn
scikit-learn
numpy
python-dotenv
google-generativeai
